+++ src/services/streams.service.ts
@@ -1,5 +1,5 @@
 import * as Endpoints from '@/api/endpoints'
-import { Stream } from '@/types'
+import { Stream, StreamsWithIncidentsParams } from '@/types'
 import ApiClient from './api.service'
 
 export async function getStreams (projects?: string[], keyword?: string): Promise<{ data: Stream[], headers: any }> {
@@ -7,6 +7,19 @@ export async function getStreams (projects?: string[], keyword?: string): Promis
   try {
     const resp = await ApiClient.request<Stream[]>({
       ...Endpoints.getStreams
+    })
+    return resp
+  } catch (e) {
+    return await Promise.reject(e)
+  }
+}
+
+export async function getStreamsWithIncidents (params: StreamsWithIncidentsParams = {}): Promise<{ data: Stream[], headers: any }> {
+  try {
+    const resp = await ApiClient.request<Stream[]>({
+      method: Endpoints.getStreams.method,
+      url: `${Endpoints.getStreams.url}/incidents`,
+      config: { params: params }
     })
     return resp
   } catch (e) {