+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 * **feature:** Display a list grouped by guardian on the Incidents page ([#47](https://github.com/rfcx/incident-center/issues/47))
 * **feature:** There are new filters on the Icidents page ([#39](https://github.com/rfcx/incident-center/issues/39))
 * **bug:** Fixed favicon issue on the Guardian Dashboard ([#44](https://github.com/rfcx/incident-center/issues/44))
+* **bug:** Use a new guardians with incidents endpoint ([#53](https://github.com/rfcx/incident-center/issues/53))
 
 ## 0.0.4 (2021-12-23)